Minor fixes and additions.
Fixed GameTime TotalTime field (it used to be the DeltaTime).
Added MathHelper.cs to aid in math functions. In this case a clamping function.
Removed the outline from BoxColliderComponent.
Removed Paused from GameScene. It didn't do anything.
Removed MenuScene. It was being used incorrectly.
Added the license for the resources.